+++ lionagi/core/generic/abc/__init__.py
@@ -15,7 +15,6 @@ from ._concepts import (
     Ordering,
     Condition,
     Actionable,
-    Workable,
     Relatable,
     Rule,
     Progressable,
@@ -30,7 +29,6 @@ __all__ = [
     "Ordering",
     "Condition",
     "Actionable",
-    "Workable",
     "Component",
     "Rule",
     "LionIDable",
