+++ lionagi/core/message/instruction.py
@@ -55,7 +55,9 @@ class Instruction(RoledMessage):
             self.content.update(context)
 
         if requested_fields:
-            self.content["requested_fields"] = self._format_output_field(requested_fields)
+            self.content["requested_fields"] = self._format_output_field(
+                requested_fields
+            )
 
     @staticmethod
     def _format_output_field(requested_fields):
