+++ lionagi/core/report/form.py
@@ -1,38 +1,35 @@
-from pydantic import Field
-
-# from lionagi import logging as _logging
-from lionagi.core.generic import BaseComponent
-from lionagi.experimental.report.util import get_input_output_fields, system_fields
+import contextlib
+from ..generic.abc import Component, Field
+from .util import get_input_output_fields, system_fields
 
 
-class Form(BaseComponent):
+class Form(Component):
 
     assignment: str = Field(..., examples=["input1, input2 -> output"])
-
     input_fields: list[str] = Field(default_factory=list)
     requested_fields: list[str] = Field(default_factory=list)
 
     def __init__(self, **kwargs):
-        """
-        at initialization, all relevant fields if not already provided, are set to None,
-        not every field is required to be filled, nor required to be declared at initialization
-        """
+
         super().__init__(**kwargs)
-        self.input_fields, self.requested_fields = get_input_output_fields(self.assignment)
+        self.input_fields, self.requested_fields = get_input_output_fields(
+            self.assignment
+        )
         for i in self.input_fields + self.requested_fields:
             if i not in self.model_fields:
                 self._add_field(i, value=None)
 
     @property
+    def filled(self):
+        with contextlib.suppress(ValueError):
+            return self._is_filled()
+        return False
+
+    @property
     def workable(self):
-        if self.filled:
-            return False
-
-        for i in self.input_fields:
-            if not getattr(self, i, None):
-                return False
-
-        return True
+        with contextlib.suppress(ValueError):
+            return self._is_workable()
+        return False
 
     @property
     def work_fields(self):
@@ -42,10 +39,6 @@ class Form(BaseComponent):
             for k, v in dict_.items()
             if k not in system_fields and k in self.input_fields + self.requested_fields
         }
-
-    @property
-    def filled(self):
-        return all([value is not None for _, value in self.work_fields.items()])
 
     def fill(self, form: "Form" = None, **kwargs):
         """
@@ -62,3 +55,19 @@ class Form(BaseComponent):
             if k not in self.work_fields:
                 raise ValueError(f"Field {k} is not a valid work field")
             setattr(self, k, v)
+
+    def _is_workable(self):
+        if self.filled:
+            raise ValueError("Form is already filled, cannot be worked on again")
+
+        for i in self.input_fields:
+            if not getattr(self, i, None):
+                raise ValueError(f"Required field {i} is not provided")
+
+        return True
+
+    def _is_filled(self):
+        for k, value in self.work_fields.items():
+            if value is None:
+                raise ValueError(f"Field {k} is not filled")
+        return True
