+++ lionagi/core/report/util.py
@@ -1,39 +1,23 @@
 from lionagi.libs import convert
 
+
 system_fields = [
-    "id_",
-    "node_id",
-    "meta",
-    "metadata",
+    "ln_id",
     "timestamp",
+    "metadata",
+    "meta",
+    "extra_fields",
     "content",
+    "created",
+    "form",
+    "report",
+    "work",
     "assignment",
     "assignments",
-    "task",
-    "template_name",
-    "version",
-    "description",
-    "in_validation_kwargs",
-    "out_validation_kwargs",
-    "fix_input",
-    "fix_output",
     "input_fields",
     "requested_fields",
-    "choices",
-    "prompt_fields",
-    "prompt_fields_annotation",
-    "instruction_context",
     "instruction",
-    "instruction_output_fields",
-    "inputs",
-    "outputs",
-    "process",
-    "_validate_field",
-    "_process_input",
-    "_process_response",
-    "_validate_field_choices",
-    "_validate_input_choices",
-    "_validate_output_choices",
+    "system",
 ]
