+++ lionagi/core/work/work_function.py
@@ -4,22 +4,29 @@ from lionagi.libs import func_call
 from functools import wraps
 from pydantic import Field
 
-from lionagi.core.generic import BaseComponent
-
-from .schema import Work, WorkLog
+from ..generic.abc import Component
 
 
-class WorkFunction:
+from .schema import Work
+from .worklog import WorkLog
+
+
+class WorkFunction(Component):
+
+    assignment: str = Field(...)
+    function: Callable = Field(...)
+    retry_kwargs: dict = Field(None)
+    guidance: str = Field(None)
+    worklog: WorkLog = Field(None)
 
     def __init__(
-        self, assignment, function, retry_kwargs=None, instruction=None, capacity=5
+        self, assignment, function, retry_kwargs=None, guidance=None, capacity=10
     ):
-
         self.assignment = assignment
         self.function = function
         self.retry_kwargs = retry_kwargs or {}
-        self.instruction = instruction or function.__doc__
-        self.worklog = WorkLog(capacity=capacity)
+        self.worklog = WorkLog(capacity)
+        self.guidance = guidance or self.function.__doc__
 
     @property
     def name(self):
@@ -28,37 +35,3 @@ class WorkFunction:
     async def perform(self, *args, **kwargs):
         kwargs = {**self.retry_kwargs, **kwargs}
         return await func_call.rcall(self.function, *args, **kwargs)
-
-    async def process(self, refresh_time=1):
-        await self.worklog.process(refresh_time=refresh_time)
-
-    async def stop(self):
-        await self.worklog.queue.stop()
-
-
-def work(assignment, capacity=5):
-    def decorator(func):
-        @wraps(func)
-        async def wrapper(self, *args, retry_kwargs=None, instruction=None, **kwargs):
-            if getattr(self, "work_functions", None) is None:
-                self.work_functions = {}
-
-            if func.__name__ not in self.work_functions:
-                self.work_functions[func.__name__] = WorkFunction(
-                    assignment=assignment,
-                    function=func,
-                    retry_kwargs=retry_kwargs or {},
-                    instruction=instruction or func.__doc__,
-                    capacity=capacity,
-                )
-
-            work_func: WorkFunction = self.work_functions[func.__name__]
-            task = asyncio.create_task(work_func.perform(*args, **kwargs))
-            work = Work(async_task=task)
-            work_func: WorkFunction = self.work_functions[func.__name__]
-            await work_func.worklog.append(work)
-            return True
-
-        return wrapper
-
-    return decorator
