+++ lionagi/core/work/worker.py
@@ -1,7 +1,12 @@
+from functools import wraps
+from typing import Any, Callable, Tuple
+
 from abc import ABC, abstractmethod
 from lionagi import logging as _logging
 from .work_function import WorkFunction
 import asyncio
+from .schema import Work
+from .worklog import WorkLog
 
 
 class Worker(ABC):
@@ -27,17 +32,76 @@ class Worker(ABC):
 
         if len(non_stopped_) > 0:
             _logging.error(f"Could not stop worklogs: {non_stopped_}")
-
         _logging.info(f"Stopped worker {self.name}")
 
-    async def process(self, refresh_time=1):
-        while not self.stopped:
-            tasks = [
-                asyncio.create_task(func.process(refresh_time=refresh_time))
-                for func in self.work_functions.values()
-            ]
-            await asyncio.wait(tasks)
-            await asyncio.sleep(refresh_time)
+
+    # TODO: Implement process method
+
+    # async def process(self, refresh_time=1):
+    #     while not self.stopped:
+    #         tasks = [
+    #             asyncio.create_task(func.process(refresh_time=refresh_time))
+    #             for func in self.work_functions.values()
+    #         ]
+    #         await asyncio.wait(tasks)
+    #         await asyncio.sleep(refresh_time)
+
+    async def _wrapper(
+        self,
+        *args,
+        func=None,
+        assignment=None,
+        capacity=None,
+        retry_kwargs=None,
+        guidance=None,
+        **kwargs,
+    ):
+        if getattr(self, "work_functions", None) is None:
+            self.work_functions = {}
+
+        if func.__name__ not in self.work_functions:
+            self.work_functions[func.__name__] = WorkFunction(
+                assignment=assignment,
+                function=func,
+                retry_kwargs=retry_kwargs or {},
+                guidance=guidance or func.__doc__,
+                capacity=capacity,
+            )
+
+        work_func: WorkFunction = self.work_functions[func.__name__]
+        task = asyncio.create_task(work_func.perform(*args, **kwargs))
+        work = Work(async_task=task)
+        work_func: WorkFunction = self.work_functions[func.__name__]
+        await work_func.worklog.append(work)
+        return True
+
+
+def work(assignment, capacity=5):
+    def decorator(func):
+        @wraps(func)
+        async def wrapper(
+            self: Worker,
+            *args,
+            func=func,
+            assignment=assignment,
+            capacity=capacity,
+            retry_kwargs=None,
+            guidance=None,
+            **kwargs,
+        ):
+            return await self._wrapper(
+                *args,
+                func=func,
+                assignment=assignment,
+                capacity=capacity,
+                retry_kwargs=retry_kwargs,
+                guidance=guidance,
+                **kwargs,
+            )
+
+        return wrapper
+
+    return decorator
 
 
 # # Example
